DAY3/Job_Portal_Api/app/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
psycopg2-binary
python-dotenv
pydantic
passlib[bcrypt]
python-jose
pytest
httpx
pytest-asyncio
alembic